Expose GetTransactionsType (#399)
Expose GetTransactionsType, getTransaction and getTransactionsCount (#399)
