+++ src/apiCalls/transactions/getTransactions.types.ts
@@ -1,0 +1,12 @@
+export interface GetTransactionsType {
+  apiAddress: string;
+  apiTimeout?: string | number;
+  sender?: string;
+  receiver?: string;
+  page?: number;
+  transactionSize?: number;
+  after?: number;
+  condition?: 'should' | 'must';
+  before?: number;
+  withScResults?: boolean;
+}
