+++ src/apiCalls/transactions/getTransactionsCount.ts
@@ -1,10 +1,7 @@
 import axios from 'axios';
 import { TRANSACTIONS_COUNT_ENDPOINT } from 'apiCalls/endpoints';
-import {
-  getTimeout,
-  getTransactionsParams,
-  GetTransactionsType
-} from './helpers';
+import { GetTransactionsType } from './getTransactions.types';
+import { getTimeout, getTransactionsParams } from './helpers';
 
 export const getTransactionsCount = (props: GetTransactionsType) => {
   const params = getTransactionsParams(props);
