+++ src/apiCalls/transactions/helpers.ts
@@ -1,15 +1,4 @@
-export interface GetTransactionsType {
-  apiAddress: string;
-  apiTimeout?: string | number;
-  sender?: string;
-  receiver?: string;
-  page?: number;
-  transactionSize?: number;
-  after?: number;
-  condition?: 'should' | 'must';
-  before?: number;
-  withScResults?: boolean;
-}
+import { GetTransactionsType } from './getTransactions.types';
 
 export const getTimeout = (apiTimeout?: string | number) =>
   apiTimeout ? { timeout: parseInt(String(apiTimeout)) } : {};
