+++ src/apiCalls/transactions/index.ts
@@ -1,3 +1,6 @@
 export * from './sendSignedTransactions';
 export * from './getTransactionsByHashes';
 export * from './getTransactions';
+export * from './getTransactionsCount';
+export * from './getTransaction';
+export * from './getTransactions.types';
